[+] fix issue where WalletConnect was initialising on backgroun on the login button display regardless of the provider [+] move components to separate files to follow the same logic as on Ledger
