+++ src/UI/walletConnect/WalletConnectConnectionStatus/index.ts
