+++ src/UI/walletConnect/WalletConnectLoginButton/index.ts
@@ -1,1 +1,2 @@
 export * from './WalletConnectLoginButton';
+export * from './types';
