+++ src/UI/walletConnect/WalletConnectLoginContainer/PairingList/ImageWithFallback.tsx
