+++ src/UI/walletConnect/WalletConnectLoginContainer/PairingList/index.ts
@@ -1,0 +1,1 @@
+export * from './PairingList';
