+++ src/UI/walletConnect/WalletConnectLoginContainer/WalletConnectLoginContent/index.ts
@@ -1,0 +1,1 @@
+export * from './WalletConnectLoginContent';
