+++ src/UI/walletConnect/WalletConnectLoginContainer/index.ts
@@ -1,0 +1,2 @@
+export * from './WalletConnectLoginContainer';
+export * from './types';
